+++ CHANGELOG.md
@@ -4,6 +4,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 
 ### Unreleased
 
+- fix(deps): update haraka-plugin-dkim to 1.2.0 to match new addresses handling #3564
 - fix(deps): update haraka-plugin-rspamd to 1.6.0 to match new addresses handling #3564
 
 ### [3.2.0] - 2026-05-NN