+++ CHANGELOG.md
@@ -5,16 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 ### Unreleased
 
 - dep(eslint): update @haraka/eslint-config to v3 and fix surfaced warnings
-
+- dep(many): bump dependencies and plugins to latest; release 3.3.0
 - refactor(auth_proxy): parse target host via net_utils.endpoint
 - refactor: move endpoint to haraka-net-utils
 - refactor: move HostPool to haraka-net-utils
 - refactor: move LineSocket to haraka-net-utils
-
 - fix(deps): update haraka-plugin-dkim to 1.2.0 to match new addresses handling #3564
 - fix(deps): update haraka-plugin-rspamd to 1.6.0 to match new addresses handling #3564
 
-### [3.2.0] - 2026-05-NN
+### [3.2.0] - 2026-05-24
 
 - fix(status): merge worker status into summary #3574
 - dep: replace address-rfc282{1,2} with @haraka/email-address #3566